pitchmyob/requirements/base.txt
boto==2.45.0
boto3==1.4.3
Django==1.10.5
django-extra-fields==0.9
django-filter==1.0.1
django-model-utils==2.6.1
django-storages==1.5.2
djangorestframework==3.5.3
-e git+https://github.com/manosim/django-rest-framework-docs@master#egg=drfdocs
Pillow==4.0.0
psycopg2==2.6.2
pytz==2016.10
Unipath==1.1
pynamodb==2.0.3
raven==5.32.0
